requirements.txt
streamlit==1.28.0
pandas==2.1.1
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.0.1
google-generativeai==0.3.2
langchain==0.0.335
langchain-community==0.0.9
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.16.3 